JAMES-2286 move MimeMessageBuilder to core
